backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pyodbc
pydantic[email]
python-jose[cryptography]
passlib
alembic
python-multipart
python-dotenv
aiofiles
mypy
pytest
httpx
pytest-asyncio
